+++ src/ui/MainWindow.h
@@ -8,7 +8,7 @@
 #include <QtWidgets/QMainWindow>
 #include <QtWidgets/QPushButton>
 #include <QSqlError>
-#include "ExpenditureView.h"
+#include "ui/table/ExpenditureView.h"
 
 class MainWindow : public QWidget
 {
