+++ src/ui/dialogs/AddCategoryDialog.cpp
