+++ src/ui/table/ExpenditureView.cpp
