+++ src/ui/table/ExpenditureView.h
